# server/swagger.yaml
...
openapi: 3.0.0
info:
  title: ldrp
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /professor/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - professor
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - professor
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/professor"
    post:
      summary: CRUD ACTION update
      tags:
        - professor
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/professor"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/professor"
  /professor/findBy_batch/{key}:
    get:
      summary: CRUD ACTION findBy_batch
      tags:
        - professor
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/professor"
  /professor/findBy_student/{key}:
    get:
      summary: CRUD ACTION findBy_student
      tags:
        - professor
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/professor"
  /professor/findBy_subject/{key}:
    get:
      summary: CRUD ACTION findBy_subject
      tags:
        - professor
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/professor"
    
  #-----  result  -----
...
  /result/{id}:
    delete:
      summary: CRUD ACTION delete
      tags:
        - result
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      summary: CRUD ACTION get
      tags:
        - result
      parameters:
        - name: id
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/result"
    post:
      summary: CRUD ACTION update
      tags:
        - result
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/result"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/result"
  /result/findBybranch/{key}:
    get:
      summary: CRUD ACTION findBybranch
      tags:
        - result
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/result"
  /result/findByprofessor/{key}:
    get:
      summary: CRUD ACTION findByprofessor
      tags:
        - result
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/result"
  /result/findBysem/{key}:
    get:
      summary: CRUD ACTION findBysem
      tags:
        - result
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/result"
  /result/findBystudent/{key}:
    get:
      summary: CRUD ACTION findBystudent
      tags:
        - result
      parameters:
        - name: key
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/result"
    
  #-----  semester  -----
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    professor:
      properties:
        _id:
          type: string
        address:
          type: string
        contact_no:
          type: number
        email:
          type: string
        f_name:
          type: string
        l_name:
          type: string
      required:
        -  _id
    result:
      properties:
        _id:
          type: string
        mark:
          type: number
      required:
        -  _id
...